requirements.txt
# crosss fade 
pydub
ffmpeg-python

# beat matching 
librosa
numpy
scipy

#frontend
streamlit

# software
fastapi
uvicorn[standard]
python-multipart
Pillow
requests
aiofiles

# experiment notebook
ipykernel

matplotlib